feat(loading_state_handler): Add LoadingStateHandlerController class
The changes introduce the `LoadingStateHandlerController` class, which is a
controller for the `LoadingStateHandlerWidget`. This controller provides a
simple way to set the state of the widget, without the need to use a
`StatefulWidget` or trigger a rebuild. The controller can be used to set the
state of the widget to `CurrentStateEnum.normal`, `CurrentStateEnum.loading`,
`CurrentStateEnum.data`, `CurrentStateEnum.empty`, or `CurrentStateEnum.error`.